+++ nugraph/models/nugraph3/decoders/__init__.py
@@ -1,0 +1,6 @@
+"""NuGraph3 decoders"""
+from .filter import FilterDecoder
+from .semantic import SemanticDecoder
+from .event import EventDecoder
+from .vertex import VertexDecoder
+from .instance import InstanceDecoder
